Implements F24 amendment
